Javascript and Stylesheet order fix
According to #10 there's an error on loada orders. This commit aims to solve CSS order issue and places the Javascript files at the bottom of <body> tag.

Further improvement on JS loading could be achieved by using defer and async option from the addScript function. This should be a theme option for admins within the backend.